API

OptMimiDICE2016R2 is an expanded version of MimiDICE2016R2. This means that every piece of code that works with MimiDICE should also work with OptMimiDICE - it is backwards-compatible. The introduced changes are merely additions that concern the optimisation of the DICE model.

Getting the model

The function OptMimiDICE2016R2.get_model is essentially unchanged, but listed here because it is central to using OptMimiDICE2016R2.

Main.OptMimiDICE2016R2.get_model — Function

get_model -> m::Model

Gets model as in standard Mimi API.

Optimising the model

The two entirely new functions are listed below. optimise_model is exported by OptMimiDICE2016R2 and can hence be called from outside the module. construct_objective is called by optimise_model and is not exported.

optimise_model — Function

optimise_model(m::Model=get_model(); kwargs) -> (m::Model, diagnostic::Dict)

Optimise DICE2016R2 model instance m and return the optimised and updated model together with diagnostic optimisation output.

The model instance m is not a mandatory argument. In case it is not provided, the function will use a newly constructed model from OptMimiDICE2016R2.get_model. It is worth manually passing a model instance if one wishes to optimise a modified version of DICE, e.g. with updated parameters or updated components.

Keyword arguments:

n_objectives::Int=length(model_years): number of objectives, which corresponds to the number of time steps in the model

stop_time::Int=640: time in seconds after which optimisation routine stops, passed to NLopt.ftol_rel!

tolerance::Float64=1e-6: tolerance requirement passed to NLopt.ftol_rel!

optimization_algorithm::Symbol=:LN_SBPLX: algorithm passed to NLopt.ftol_rel!

backup_timesteps::Int=0: amount of time steps in model's time dimension before optimisation sets in

Notes

This version of DICE only allows for NETs after 2150, and it immediately allows for a 20% emissions reduction rate (:MIU=1.2) in 2155. This constraint can be modified by changing the first half of the upper_bound vector.

The second return value is purely for diagnostic purposes and comes directly from the NLopt optimisation. In normal usage, it can be ignored.

See also construct_objective.

construct_objective — Function

construct_objective(m::Model, optimised_mitigation::Array{Float64,1}, backup_timesteps::Int=0, n_objectives::Int=length(model_years)) -> m[:welfare, :UTILITY]

Updates emissions control rate :MIU and savings rate :S in model m and returns the resulting utility vector. This function is called by optimise_model. optimised_mitigation_savings is a vector of :MIU and :S values that is being optimised.

backup_timesteps gives the amount of timesteps that are part of the model's time dimension without being optimised. For example, the FaIR climate module runs since 1765, but DICE only optimises starting in 2015. n_objectives gives the amount of timesteps that are being optimised, such that there are n_objectives for :MIU and n_objectives for :S.

See also optimise_model.

Index

Main.OptMimiDICE2016R2.get_model

construct_objective

optimise_model